Performance improvement for alpha_blend (3-4x)
